requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.3.7
python-dotenv==1.0.0
slack_sdk